Get/set the Hubspot secret API key

Get/set the Hubspot secret API key using the keyring package. For a more secure authorization method, without a daily limit on calls, see hubspot_token_create()

hubspot_key_get()

hubspot_key_set(key = NULL)

Arguments

key
API key. If not provided via this argument, the key will have to be provided interactively.

Value

The key (a string)

Details

To know how to get your Hubspot API key, refer to the corresponding section of Hubspot API docs. Hubspot recommends you to re-generate your API key every six months.

The API key should be set using the hubspot_key_set() function, that relies on the keyring package that stores the key in the system default credential store (see keyring docs).

# Either input the key as parameter
hubspot_key_set("yourapikey")

# Or just call the function,
# the key will be asked interactively
hubspot_key_set()

If no API key is set, the “demo” key is used and a message is printed to the screen. As noted earlier, this won’t give you access to your own Hubspot data.

You can check the daily usage related to the API key via hubspot_api_usage(). Calls in the package are rate limited (100 requests every 10 seconds) but there’s also a daily limit on calls.

hubspot_api_usage()

For a more secure authorization method, without a daily limit on calls, see hubspot_token_create().

Authorization for Hubspot APIs

The Hubspot API accepts authorization via

an API key,

OAuth 2.0.

OAuth 2.0 is the recommended method. However, this package supports both.

Note that if you do nothing the package will use the “demo” API token but this won’t give you access to your own Hubspot data. So you’ll need to spend a little time on setup:

For rapid prototyping key, use a Hubspot API key, see hubspot_key_set().

For more secure use, without a daily limit on API calls, see hubspot_token_create() to create a Hubspot authorization token (OAuth 2.0).

If you have both saved an API key via hubspot_key_set() and a token via hubspot_token_create(), priority will be given to using the OAuth 2.0 token. If you don’t want that, explicitely pass NULL as value for the token_path argument of all functions.

See also

Examples

hubspot_key_get()

#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables

#> Could not find Hubspot API key, will use the 'demo' key.
#> Please run hubspot_key_set() to set your API key.

#> [1] "demo"

# \dontrun{
hubspot_key_set("yourapikey")

#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables

#> Key saved for the service 'hubspot'

# }