Create/get Hubspot authorization token.

Functions related to creating or retrieving an OAuth 2.0 token for the Hubspot API.

default_ld_hubspot_app()

hubspot_token_create(
  app_info = default_ld_hubspot_app(),
  set_renv = TRUE,
  token_path = NULL
)

hubspot_token_get()

Arguments

app_info
A named list with client_secret, client_id, app_id and scopes.

set_renv
Logical indicating whether to save the created token as the default environment hubspot token variable. Defaults to TRUE, meaning the token is saved to user's home directory as either the user provided path, or ".hubspot_token.rds" (or, if that already exists, then .hubspot_token1.rds or .hubspot_token2.rds, etc.) and the path to the token to said token is then set in the user's .Renviron file and re- read to start being used in current active session.

token_path
Path where to save the token. If set_renv is FALSE, this is ignored.

Value

Named list

Either NULL or the path in which the token is saved.

Details

We expect you to not need to bring your own app (BYOA), but the app_info parameter of hubspot_token_create() ensures you can do that if you want (for instance to define an app with only the scope needed for the endpoints you use as opposed to all the endpoints the hubspot package supports).

To BYOA,

Read and execute Hubspot’s docs about how to create an app.

Define the app_info list.

my_very_own_app <- list(client_secret = <CLIENTSECRET>,
       client_id = <CLIENTID>,
       app_id = <APPID>,
       scope = c(<"scope1", "scope2", etc>))

Pass the app info to hubspot_token_create()

hubspot_token_create(app_info = my_very_own_app)

To distribute your app across your projects or across team members, you could make it a function in an internal package, or use environment variables.

Authorization for Hubspot APIs

The Hubspot API accepts authorization via

an API key,

OAuth 2.0.

OAuth 2.0 is the recommended method. However, this package supports both.

Note that if you do nothing the package will use the “demo” API token but this won’t give you access to your own Hubspot data. So you’ll need to spend a little time on setup:

For rapid prototyping key, use a Hubspot API key, see hubspot_key_set().

For more secure use, without a daily limit on API calls, see hubspot_token_create() to create a Hubspot authorization token (OAuth 2.0).

If you have both saved an API key via hubspot_key_set() and a token via hubspot_token_create(), priority will be given to using the OAuth 2.0 token. If you don’t want that, explicitely pass NULL as value for the token_path argument of all functions.

See also

Examples

# \dontrun{
hubspot_token_create()

#> httpuv not installed, defaulting to out-of-band authentication

#> Error: Can only use oob authentication in an interactive session

# }
hubspot_token_get()

#> NULL